+++ modules/client/CHANGELOG.md
@@ -22,6 +22,7 @@ TEMPLATE:
 ### Added
 
 - Add prepareUpdate function to all clients
+- Addd check for dao update proposal
 
 ### Changed